Move getPreferredIdentifier() into github.js
It assumes a GitHub-structured set of data about the vulnerability, so
it really is a GitHub-specific method.
