+++ src/csv.js
@@ -1,6 +1,6 @@
 const { format } = require('@fast-csv/format')
 const fs = require("fs")
-const { getPreferredIdentifier } = require('./index')
+const { getPreferredIdentifier } = require('./github')
 
 /**
  * Write the given vulnerabilities data to a CSV file.
