+++ src/github.js
@@ -59,6 +59,16 @@ module.exports.getDependenciesOfRepo = async (token, repo) => {
   return composer.getDependenciesFromComposerLock(contents)
 }
 
+module.exports.getPreferredIdentifier = vulnerability => {
+  const advisory = vulnerability.advisory || {}
+  let identifiers = advisory.identifiers || []
+  const cveIdentifiers = identifiers.filter(i => i.type === 'CVE')
+  if (cveIdentifiers.length > 0) {
+    identifiers = cveIdentifiers
+  }
+  return identifiers.map(i => i.value).join(', ')
+}
+
 const getSecurityVulnerabilitiesForPackage = async (token, packageName) => {
   console.log(`Calling GitHub API for security vulnerabilities for ${packageName}`)
   const data = await graphql({
