+++ src/index.js
@@ -2,16 +2,6 @@ const composer = require('./composer')
 const github = require('./github')
 const semverSatisfies = require('semver/functions/satisfies')
 
-module.exports.getPreferredIdentifier = vulnerability => {
-  const advisory = vulnerability.advisory || {}
-  let identifiers = advisory.identifiers || []
-  const cveIdentifiers = identifiers.filter(i => i.type === 'CVE')
-  if (cveIdentifiers.length > 0) {
-    identifiers = cveIdentifiers
-  }
-  return identifiers.map(i => i.value).join(', ')
-}
-
 const getSecurityVulnerabilitiesForDependency = async (token, dependency) => {
   const vulnerabilities = await github.getSecurityVulnerabilitiesForPackageUsingCache(
     token,
